+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 - [commands] Add additional CommandInfo as such `Description` & `Usage` & `LongDescription`
 - [commands] Fixes duplication of tab completion options
 - [commands] Add ignore case for map based parser
+- [commands] Add dynamic parser for method/class
 - [common] Add Rotation and Transform classes
-- [docs] Add docs talking about region box, transform, region and rotation.
+- [docs] Add docs about region box, transform, region and rotation.
 
 ## v0.0.29
 - [commands] Improve location parser
@@ -14,4 +15,4 @@
 - [commands] optional or invalid arguments
 - [spigot] visibility manager api
 - [commands] Polymorphic command declaration
-- [docs] Introduce new docs based on Writerside
+- Introduce new docs based on Writerside